# Pindrop-Production/lib/api-zod/src/generated/types/resolveInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResolveInput {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  url: string;
}