# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFormat } from './mediaFormat';
import type { MediaType } from './mediaType';

export interface Media {
  type: MediaType;
  format: MediaFormat;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  size?: number | null;
  download_url: string;
}